COMMIT BY GCC 2021

+++ Question8/Question8.cpp
@@ -1,9 +1,9 @@
 #include <bits/stdc++.h>
 
 using namespace std;
-int pos, max_profit, cur_profit, j;
+long long pos, max_profit, cur_profit, j;
 
-int calculateMaximizedReturns(int n, int k, int d, int m, vector<int> returns) {
+long long calculateMaximizedReturns(int n, int k, int d, long long m, vector<long long> returns) {
 	vector<int> status(n, 1);
 	for (int a = 0; a < k; ++a) {
 		pos = -1;
@@ -57,61 +57,61 @@ int calculateMaximizedReturns(int n, int k, int d, int m, vector<int> returns) {
 }
 
 vector<string> split(const string& str, char delim) {
-    vector<string> strings;
-    size_t start;
-    size_t end = 0;
-    while ((start = str.find_first_not_of(delim, end)) != string::npos) {
-        end = str.find(delim, start);
-        strings.push_back(str.substr(start, end - start));
-    }
-    return strings;
+	vector<string> strings;
+	size_t start;
+	size_t end = 0;
+	while ((start = str.find_first_not_of(delim, end)) != string::npos) {
+		end = str.find(delim, start);
+		strings.push_back(str.substr(start, end - start));
+	}
+	return strings;
 }
 
-vector<int> splitStringToInt(const string& str, char delim) {
-    vector<int> strings;
-    size_t start;
-    size_t end = 0;
-    while ((start = str.find_first_not_of(delim, end)) != string::npos) {
-        end = str.find(delim, start);
-        strings.push_back(stoi(str.substr(start, end - start)));
-    }
-    return strings;
+vector<long long> splitStringToInt(const string& str, char delim) {
+	vector<long long> strings;
+	size_t start;
+	size_t end = 0;
+	while ((start = str.find_first_not_of(delim, end)) != string::npos) {
+		end = str.find(delim, start);
+		strings.push_back(stoi(str.substr(start, end - start)));
+	}
+	return strings;
 }
 
 void printVector(vector<int> vec) {
-    for (vector<int>::const_iterator i = vec.begin(); i != vec.end(); ++i) {
-        cout << *i << ' ';
-    }
-    cout << endl;
+	for (vector<int>::const_iterator i = vec.begin(); i != vec.end(); ++i) {
+		cout << *i << ' ';
+	}
+	cout << endl;
 }
 
 void printVector(vector<string> vec) {
-    for (vector<string>::const_iterator i = vec.begin(); i != vec.end(); ++i) {
-        cout << *i << ' ';
-    }
-    cout << endl;
+	for (vector<string>::const_iterator i = vec.begin(); i != vec.end(); ++i) {
+		cout << *i << ' ';
+	}
+	cout << endl;
 }
 
 int main() {
-    string firstLine;
-    getline(cin, firstLine);
+	string firstLine;
+	getline(cin, firstLine);
 
-    vector<int> firstLineVec = splitStringToInt(firstLine, ' ');
-    int n = firstLineVec[0];
-    int k = firstLineVec[1];
-    int d = firstLineVec[2];
-    int m = firstLineVec[3];
+	vector<long long> firstLineVec = splitStringToInt(firstLine, ' ');
+	int n = firstLineVec[0];
+	int k = firstLineVec[1];
+	int d = firstLineVec[2];
+	int m = firstLineVec[3];
 
-    string returns;
-    getline(cin, returns);
+	string returns;
+	getline(cin, returns);
 
-    vector<int> returnsVec = splitStringToInt(returns, ' ');
+	vector<long long> returnsVec = splitStringToInt(returns, ' ');
 
-    int result = calculateMaximizedReturns(n, k, d, m, returnsVec);
+	long long result = calculateMaximizedReturns(n, k, d, m, returnsVec);
 
-    // Do not remove below line
-    cout << result << "\n";
-    // Do not print anything after this line
+	// Do not remove below line
+	cout << result << "\n";
+	// Do not print anything after this line
 
-    return 0;
+	return 0;
 }
